[IR][Arith] Use BigInt-backed integer immediates and wide coefficients (#20371)
Store shared integer immediate values in the existing `ffi::BigInt`,
preserving declared signedness and width checks. Full-range uint64
constants become ordinary `IntImm` nodes, and the `large_uint_imm`
builtin is removed. Constant construction and folding use shared
primitive helpers. Construction, transport, printing, and backend
emission preserve the exact value or reject unsupported values and
types.

Constant predicates recognize every IntImm, while consumers with native
integer domains perform their own checked conversions. Immediate
patterns capture the original typed node, with exact BigInt conditions
and expression intermediates. Canonical and product coefficients retain
exact values. Canonical split factors and their divisor/modulus
arithmetic use BigInt, with an optional upper factor distinguishing
finite moduli from no modulus. Signed rewrites preserve rounding, and
IntImm construction enforces the declared integer range for synthesized
constants. Typed folding preserves integer overflow contracts. Modular
analysis and constant bounds keep their int64 domains and conservatively
handle wider immediates. Canonical index-type eligibility, dimensions,
counters and backend interfaces retain their existing limits.

+++ src/backend/trn/codegen/codegen_trn.cc
@@ -248,15 +248,15 @@ void CodeGenTrainium::Dispatch_(const AllocBufferNode* op) {
           << "allocated_addr[0] must be a constant integer, got: " << addr[0];
       TVM_FFI_ICHECK(addr[1]->IsInstance<IntImmNode>())
           << "allocated_addr[1] must be a constant integer, got: " << addr[1];
-      int64_t base_bank = addr[0].as_or_throw<IntImm>()->value;
-      int64_t base_addr = addr[1].as_or_throw<IntImm>()->value;
+      int64_t base_bank = static_cast<int64_t>(addr[0].as_or_throw<IntImm>()->value);
+      int64_t base_addr = static_cast<int64_t>(addr[1].as_or_throw<IntImm>()->value);
       stream << "ncc.psum.mod_alloc(base_bank=" << base_bank << ", base_addr=" << base_addr;
       stream << ", num_bank_tiles=(" << op->buffer->shape[0] << ",)))\n";
     } else {
       TVM_FFI_ICHECK(addr.size() == 1);
       TVM_FFI_ICHECK(addr[0]->IsInstance<IntImmNode>())
           << "allocated_addr[0] must be a constant integer, got: " << addr[0];
-      int64_t base_addr = addr[0].as_or_throw<IntImm>()->value;
+      int64_t base_addr = static_cast<int64_t>(addr[0].as_or_throw<IntImm>()->value);
       stream << "ncc.sbuf.mod_alloc(base_addr=" << base_addr << "))\n";
     }
   }
